Refactor initialization code (delete app-oclif.ts and app-common.ts)
Change-type: patch
